+++ codeHF/runtest.sh
@@ -103,6 +103,7 @@ fi
 
 # Convert AliESDs.root to AO2D.root.
 if [ $DOCONVERT -eq 1 ]; then
+  [ -f "$LISTFILESALI" ] || { echo "Converting: Error: File $LISTFILESALI does not exist."; exit 1; }
   echo -e "\nConverting... ($(cat $LISTFILESALI | wc -l) files)"
   if [ $DOQA -eq 1 ]; then
     echo "Setting MC mode ON."
@@ -110,14 +111,12 @@ if [ $DOCONVERT -eq 1 ]; then
   fi
   if [ $CONVSEP -eq 1 ]; then
     echo "Converting files separately"
-    $ENVALI bash convert_batch.sh $LISTFILESALI $LISTFILESO2 $ISMC $DEBUG # Run the batch script in the ALI environment.
-    if [ $? -ne 0 ]; then exit 1; fi # Exit if error.
+    $ENVALI bash convert_batch.sh $LISTFILESALI $LISTFILESO2 $ISMC $DEBUG || exit 1 # Run the batch script in the ALI environment.
   else
     LOGFILE="log_convert.log"
     rm -f $LOGFILE
     echo "logfile: $LOGFILE"
-    $ENVALI $CMDROOT "convertAO2D.C(\"$LISTFILESALI\", $ISMC, $NMAX)" > $LOGFILE 2>&1
-    if [ $? -ne 0 ]; then echo "Error"; exit 1; fi # Exit if error.
+    $ENVALI $CMDROOT "convertAO2D.C(\"$LISTFILESALI\", $ISMC, $NMAX)" > $LOGFILE 2>&1 || { echo "Error"; exit 1; }
     echo "$PWD/AO2D.root" > $LISTFILESO2
     rm -f $FILEOUTO2
   fi
@@ -126,6 +125,7 @@ fi
 # Perform simple QA studies with O2.
 if [ $DOQA -eq 1 ]; then
   #LOGFILE="log_o2_qa.log"
+  [ -f "$O2INPUT" ] || { echo "QA task: Error: File $O2INPUT does not exist."; exit 1; }
   echo -e "\nRunning the QA task with O2... ($(cat $O2INPUT | wc -l) files)"
   rm -f $FILEOUTO2 $FILEOUTQA
   O2ARGS="--shm-segment-size 16000000000 --configuration json://$JSON"
@@ -140,11 +140,9 @@ if [ $DOQA -eq 1 ]; then
 #!/bin/bash
 $O2EXEC
 EOF
-  #$ENVO2 bash $O2SCRIPT > $LOGFILE 2>&1 # Run the script in the O2 environment.
-  #if [ $? -ne 0 ]; then echo "Error"; exit 1; fi # Exit if error.
+  #$ENVO2 bash $O2SCRIPT > $LOGFILE 2>&1 || { echo "Error"; exit 1; } # Run the script in the O2 environment.
   #grep WARN $LOGFILE | sort -u
-  $ENVO2 bash o2_batch.sh $O2INPUT $JSON $O2SCRIPT $DEBUG # Run the batch script in the O2 environment.
-  if [ $? -ne 0 ]; then exit 1; fi # Exit if error.
+  $ENVO2 bash o2_batch.sh $O2INPUT $JSON $O2SCRIPT $DEBUG || exit 1 # Run the batch script in the O2 environment.
   rm -f $O2SCRIPT
   mv $FILEOUTO2 $FILEOUTQA
   rm -rf output_o2_qa
@@ -154,15 +152,16 @@ fi
 
 # Run the heavy-flavour tasks with AliPhysics.
 if [ $DORUN1 -eq 1 ]; then
+  [ -f "$LISTFILESALI" ] || { echo "HF tasks ALI: Error: File $LISTFILESALI does not exist."; exit 1; }
   echo -e "\nRunning the HF tasks with AliPhysics... ($(cat $LISTFILESALI | wc -l) files)"
   #$ENVALI bash ali_batch.sh $LISTFILESALI $JSON $FILEOUTALI # Run the batch script in the ALI environment.
-  $ENVALIO2 bash ali_batch.sh $LISTFILESALI $JSON $FILEOUTALI $DEBUG # Run the batch script in the ALI+O2 environment.
-  if [ $? -ne 0 ]; then exit 1; fi # Exit if error.
+  $ENVALIO2 bash ali_batch.sh $LISTFILESALI $JSON $FILEOUTALI $DEBUG || exit 1 # Run the batch script in the ALI+O2 environment.
 fi
 
 # Run the heavy-flavour tasks with O2.
 if [ $DORUN3 -eq 1 ]; then
   #LOGFILE="log_o2_hf.log"
+  [ -f "$O2INPUT" ] || { echo "HF tasks O2: Error: File $O2INPUT does not exist."; exit 1; }
   echo -e "\nRunning the HF tasks with O2... ($(cat $O2INPUT | wc -l) files)"
   rm -f $FILEOUTO2
   # Option --configuration has priority over --aod-file.
@@ -193,11 +192,9 @@ if [ $DORUN3 -eq 1 ]; then
 #!/bin/bash
 $O2EXEC
 EOF
-  #$ENVO2 bash $O2SCRIPT > $LOGFILE 2>&1 # Run the script in the O2 environment.
-  #if [ $? -ne 0 ]; then echo "Error"; exit 1; fi # Exit if error.
+  #$ENVO2 bash $O2SCRIPT > $LOGFILE 2>&1 || { echo "Error"; exit 1; } # Run the script in the O2 environment.
   #grep WARN $LOGFILE | sort -u
-  $ENVO2 bash o2_batch.sh $O2INPUT $JSON $O2SCRIPT $DEBUG # Run the batch script in the O2 environment.
-  if [ $? -ne 0 ]; then exit 1; fi # Exit if error.
+  $ENVO2 bash o2_batch.sh $O2INPUT $JSON $O2SCRIPT $DEBUG || exit 1 # Run the batch script in the O2 environment.
   rm -f $O2SCRIPT
   rm -rf output_o2_hf
   mv output_o2 output_o2_hf
@@ -214,14 +211,10 @@ if [ $DOCOMPARE -eq 1 ]; then
   echo -e "\nComparing... (logfile: $LOGFILE)"
   ok=1
   for file in "$FILEOUTALI" "$FILEOUTO2"; do
-    if [ ! -f "$file" ]; then
-      echo "Error: File $file does not exist."
-      ok=0
-    fi
+    [ -f "$file" ] || { echo "Error: File $file does not exist."; ok=0; }
   done
-  if [ $ok -ne 1 ]; then exit 1; fi
-  $ENVALI $CMDROOT "Compare.C(\"$FILEOUTO2\",\"$FILEOUTALI\", $MASS)" > $LOGFILE 2>&1
-  if [ $? -ne 0 ]; then echo "Error"; exit 1; fi # Exit if error.
+  [ $ok -ne 1 ] && exit 1
+  $ENVALI $CMDROOT "Compare.C(\"$FILEOUTO2\",\"$FILEOUTALI\", $MASS)" > $LOGFILE 2>&1 || { echo "Error"; exit 1; }
 fi
 
 echo -e "\nDone"
